[misc] - refactor: rename `Tab` component and story files to `Tabs`
 - Standardized the component and story filenames to reflect plurality (Tabs instead of Tab)
